Port eslint rules over to new eslint rule format.
See http://eslint.org/blog/2016/07/eslint-new-rule-format for a few more
details.

Fixes #1135
