+++ src/rules/javascript/only-prefs-in-defaults.js
@@ -1,0 +1,23 @@
+import { ONLY_PREFS_IN_DEFAULTS } from 'messages/javascript';
+import { getRootExpression } from 'utils';
+
+export default {
+  create(context) {
+    var filename = context.getFilename();
+
+    // This rule only applies to files in defaults/preferences
+    if (filename.indexOf('defaults/preferences/') === 0) {
+      return {
+        CallExpression: function(node) {
+          var root = getRootExpression(node);
+
+          if (root.name !== 'pref' && root.name !== 'user_pref') {
+            return context.report(node, ONLY_PREFS_IN_DEFAULTS.code);
+          }
+        },
+      };
+    } else {
+      return {};
+    }
+  },
+};
