+++ src/rules/javascript/webextension-api.js
@@ -1,0 +1,24 @@
+import { DEPRECATED_APIS, TEMPORARY_APIS } from 'const';
+import { apiToMessage } from '../../utils';
+
+export default {
+  create(context) {
+    return {
+      MemberExpression: function(node) {
+        if (node.object.object &&
+            ['chrome', 'browser'].includes(node.object.object.name)) {
+          let api = `${node.object.property.name}.${node.property.name}`;
+
+          if (DEPRECATED_APIS.includes(api)) {
+            return context.report(node, apiToMessage(api));
+          }
+
+          if (!context.settings.addonMetadata.id &&
+              TEMPORARY_APIS.includes(api)) {
+            return context.report(node, apiToMessage(api));
+          }
+        }
+      },
+    };
+  },
+};
